+++ client.go
@@ -8,7 +8,7 @@ import (
 	"time"
 )
 
-type Context struct {
+type Client struct {
 	*sql.DB
 	sync.Mutex
 	_ noCopy
@@ -20,11 +20,11 @@ type Context struct {
 	Index           int
 }
 
-func (db *Context) Query(query string, args ...any) (*Rows, error) {
+func (db *Client) Query(query string, args ...any) (*Rows, error) {
 	return db.QueryContext(context.Background(), query, args...)
 }
 
-func (db *Context) QueryBuilder(ctx context.Context, b *Builder) (*Rows, error) {
+func (db *Client) QueryBuilder(ctx context.Context, b *Builder) (*Rows, error) {
 	query, args, err := b.Build()
 	if err != nil {
 		return nil, err
@@ -33,7 +33,7 @@ func (db *Context) QueryBuilder(ctx context.Context, b *Builder) (*Rows, error) 
 	return db.QueryContext(ctx, query, args...)
 }
 
-func (db *Context) QueryContext(ctx context.Context, query string, args ...any) (*Rows, error) {
+func (db *Client) QueryContext(ctx context.Context, query string, args ...any) (*Rows, error) {
 	var rows *sql.Rows
 	var stmt *Stmt
 	var err error
@@ -59,11 +59,11 @@ func (db *Context) QueryContext(ctx context.Context, query string, args ...any) 
 	return &Rows{Rows: rows, stmt: stmt, query: query}, nil
 }
 
-func (db *Context) QueryRow(query string, args ...any) *Row {
+func (db *Client) QueryRow(query string, args ...any) *Row {
 	return db.QueryRowContext(context.Background(), query, args...)
 }
 
-func (db *Context) QueryRowBuilder(ctx context.Context, b *Builder) *Row {
+func (db *Client) QueryRowBuilder(ctx context.Context, b *Builder) *Row {
 	query, args, err := b.Build()
 	if err != nil {
 		return &Row{
@@ -75,7 +75,7 @@ func (db *Context) QueryRowBuilder(ctx context.Context, b *Builder) *Row {
 	return db.QueryRowContext(ctx, query, args...)
 }
 
-func (db *Context) QueryRowContext(ctx context.Context, query string, args ...any) *Row {
+func (db *Client) QueryRowContext(ctx context.Context, query string, args ...any) *Row {
 	var rows *sql.Rows
 	var stmt *Stmt
 	var err error
@@ -108,11 +108,11 @@ func (db *Context) QueryRowContext(ctx context.Context, query string, args ...an
 	}
 }
 
-func (db *Context) Exec(query string, args ...any) (sql.Result, error) {
+func (db *Client) Exec(query string, args ...any) (sql.Result, error) {
 	return db.ExecContext(context.Background(), query, args...)
 }
 
-func (db *Context) ExecBuilder(ctx context.Context, b *Builder) (sql.Result, error) {
+func (db *Client) ExecBuilder(ctx context.Context, b *Builder) (sql.Result, error) {
 	query, args, err := b.Build()
 	if err != nil {
 		return nil, err
@@ -121,7 +121,7 @@ func (db *Context) ExecBuilder(ctx context.Context, b *Builder) (sql.Result, err
 	return db.ExecContext(ctx, query, args...)
 }
 
-func (db *Context) ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error) {
+func (db *Client) ExecContext(ctx context.Context, query string, args ...any) (sql.Result, error) {
 	if len(args) > 0 {
 		stmt, err := db.prepareStmt(ctx, query)
 		if err != nil {
@@ -135,12 +135,12 @@ func (db *Context) ExecContext(ctx context.Context, query string, args ...any) (
 	return db.DB.ExecContext(context.Background(), query, args...)
 }
 
-func (db *Context) Begin(opts *sql.TxOptions) (*Tx, error) {
+func (db *Client) Begin(opts *sql.TxOptions) (*Tx, error) {
 	return db.BeginTx(context.TODO(), opts)
 
 }
 
-func (db *Context) BeginTx(ctx context.Context, opts *sql.TxOptions) (*Tx, error) {
+func (db *Client) BeginTx(ctx context.Context, opts *sql.TxOptions) (*Tx, error) {
 	tx, err := db.DB.BeginTx(ctx, opts)
 	if err != nil {
 		return nil, err
@@ -149,7 +149,7 @@ func (db *Context) BeginTx(ctx context.Context, opts *sql.TxOptions) (*Tx, error
 	return &Tx{Tx: tx, stmts: make(map[string]*sql.Stmt)}, nil
 }
 
-func (db *Context) Transaction(ctx context.Context, opts *sql.TxOptions, fn func(ctx context.Context, tx *Tx) error) error {
+func (db *Client) Transaction(ctx context.Context, opts *sql.TxOptions, fn func(ctx context.Context, tx *Tx) error) error {
 	tx, err := db.BeginTx(ctx, opts)
 	if err != nil {
 		return err